move serialization-related methods to serialization.py
update tests that call specific methods from this area
